feat(CiNomModal): add dataOwners state and update options
Add a new state `dataOwners` to store processed owner data and update the options in the form to use this state. This ensures that the owner names are consistently formatted before being used in the UI.

feat(GroupQR): add loading state for data

Introduce a loading component to display while the data is being fetched. This improves the user experience by providing feedback during data loading.